students-client: Generate the whole React code using an AI agent
The model used was claude-4-sonnet. The following prompt was used:

In this project, you have several directories. The goal is to create a React application within `students-client`. This is the only directory, in which you'll be creating or changing files.
In `vzor`, you have static HTML files (and a CSS file) of how the result should look like. When generating JSX, follow the HTML as closely as possible. The application has views for a list of students, for a student detail, and for creating/updating a student. In the list, there's also the option of deleting a student.
In `students-server-restapi-node/students.yaml`, there's an OpenAPI definition of a REST API around the student entity. In the directory, there's also a functioning implementation but you shouldn't need that; the OpenAPI description should suffice. Just assume this server is running at http://localhost:8080. Apart from operations regarding the student entity, there's also the /codebooks/{codebookCode} endpoint. This is used to obtain codebook values for `gender`, `house`, and `year`. You'll need these not only within the forms but also to properly show the student data because its `gender`, `house`, and `year` attributes are just codes referencing these codebook values. As you can see, codebook item values are localized. Assume there is always an English localization (under the `en` key) and use only this one localization.
The resulting React application will be created within `students-client`. There is already an empty structure, as generated by Vite. I've also added a dependency on `react-router-dom`. You are not allowed to use any additional dependencies. Using `react-router-dom`, create the following routing:
- "/" will contain the list of students,
- "/students/:id" will contain the detail of a student,
- "/students/:id/edit" will contain the update form of a student,
- "/students/create" will contain the create form of a student.