students-client: Fix working with codebooks
The conversation for this was a bit longer because I misjudged where the
problem was. I thought the AI had forgotten to implement it, however,
the problem was simply in the codebook codes not being in uppercase.

My first prompt:

====
Now, let's go screen by screen. The list of students is properly fetched from the server but for `gender`, `house`, and `year`, only the codes (which are returned by the server) are shown. Instead, appropriate codebook values (for English, i.e. `en` localization) are supposed to be shown.
====

For this, the agent wanted to actually download the codebook data from
the server, which I didn't want to allow it.

====
You shouldn't need to have the server running. Look at the OpenAPI definition. Is the description (of working with codebooks) clear enough for you? If not, don't hesitate to ask me.
====

After this, it asked me how I wanted to handle errors, which made me
check that the code regarding codebooks was mostly fine, except that it
didn't use uppercase codes.

====
Ah, I see where the problem might be. I neglected to mention that codes of the codebooks themselves are uppercase, so to get `gender`, you need to use http://localhost:8080/codebooks/GENDER. Can you update your code?
====