students-client: Add CSS for StudentList
The prompt:

I'd like to look at the styles now. I don't want to use anything sophisticated but if there are, for example, any styles specific for elements created/used only within StudentList, let's create StudentList.css, import it in the StudentList component, and let's just use "StudentList" as a class name of the top-level JSX component rendered in the StudentList component (as a sort of poor man's solution for namespaces). Is there any rule in the global CSS file that could be moved into this specific file?